Add FREM covariates to a FFEM data set

Add binarised FREM covariates to a FFEM data set.

Usage

addFREMcovariates(
  dfFFEM,
  modFile = NULL,
  covariates = NULL,
  iMiss = -99,
  includeReference = FALSE,
  imputeMissing = TRUE
)

Arguments

dfFFEM

A data.frame containing the FFEM data.

modFile

Path to a FREM model file.

covariates

A character vector of covariates to binarise.

iMiss

The integer/numeric value representing missing data (default: -99).

includeReference

Logical. If TRUE, creates a dummy column for the lowest reference level as well (default: FALSE).

imputeMissing

Logical. If TRUE (default), missing values are imputed to 0 (the reference category). If FALSE, missing values are preserved as iMiss.

Value

A data.frame with the new binarised covariate columns.

Details

This is mainly an internal utility function.

Binarised covariates will be added based either on the covariates present in a frem model (modFile) file or from the names specified in the covariates vector.

The new columns in the FFEM data set will contain the binarised versions of the polychotomous covariates named with the original covariate name appended by "_#", where "#" is the level value of the covariate. For example, for a covariate NCIL with the values 0, 1 and 2 addFREMcovariates() will add columns NCIL_1 and NCIL_2 each with the values 0 and 1 for rows where the value of NCIL is not 1/2 and 1 when the in NCIL is 1 or 2, respectively. If an ID has a missing value of NCIL, then NCIL_1 and NCIL_2 will both be 0, effectively imputing the missing NCIL with the lowest level of the NCIL covariate.

See also

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Load the base dataset
dfFFEM <- read.csv(
  system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv", package="PMXFrem")) %>% 
  filter(BLQ != 1)

modFile <- system.file("extdata/SimNeb/run31.mod", package="PMXFrem")

## Example 1: Add covariates automatically based on a FREM model file
dfFFEM_mod <- addFREMcovariates(dfFFEM, modFile = modFile)

## Example 2: Add specific polychotomous covariates (Default: Y-1 categories)
# For RACEL (levels 1, 2, 3), this will generate RACEL_3 and RACEL_2.
# The lowest valid level (1) is dropped as the reference category.
dfFFEM_standard <- addFREMcovariates(dfFFEM, covariates = "RACEL")

## Example 3: Add specific covariates AND include the reference category
# This will generate RACEL_3, RACEL_2, AND RACEL_1.
dfFFEM_full <- addFREMcovariates(dfFFEM, covariates = "RACEL", includeReference = TRUE)