Calculate conditional ETAs (ETA prim)

Collects the ETAs (parameter and covariate) from a FREM model and computes the ETA_prims.

Usage

calcEtas(
  runno = NULL,
  numNonFREMThetas,
  modName = NULL,
  numSkipOm = 0,
  idvar = "ID",
  modDevDir = NULL,
  FFEMData = NULL,
  covmodel = "linear",
  dataFile = NULL,
  parNames = NULL,
  quiet = TRUE,
  ...
)

Arguments

runno

The run number. Must be provided if modName is NULL. Will be used to create NONMEM model and results file names with the structure run#.mod and run#. NONMEM ext, phi and cov files will be assumed to be named run# followed by the appropriate suffix (where # is runno).

numNonFREMThetas

Number of thetas in the FREM model that are not FREM covariates. These need to come before the FREM covariate thetas in the FREM model file.

modName

The model name. Will be used together with modExt, lstExt, ext, phi and cov to form the NONMEM file names, for example modName.mod for the model file.

numSkipOm

Number of diagonal omegas that should not be part of the FREM calculations. Such omegas has to come before the large FREM omega block in the FREM model file.

idvar

The name of the ID column,

modDevDir

The directory where the files reside. Default is the current directory, i.e. ".".

FFEMData

An FFEMData object as obtained with the function createFFEMdata. If NULL, the object will be created internally, requiring dataFile, parNames, and other arguments for createFFEMdata to be provided.

covmodel

A character string indicating if the covariate models were implemented linearly (additatively) in the frem model or not. Default is "linear".

dataFile

The name of the data file used in the base model, i.e. the original data file, or a data.frame with the same data.

parNames

Names of the parameters. Only used in the STDOUT when quiet =FALSE or in eqFile if it is non-blank. Defaults to Par folowed by a number.

quiet

If FALSE, will print the FFEM model + associated $OMEGA BLOCK to STDOUT.

...

Additional arguments passed on to createFFEMdata when FFEMData is NULL.

Value

A data.frame with one row per subject, containing the individual parameter estimates. The data frame includes the following columns:

ID: The subject identifier.

ETA_PRIM*: One column for each parameter's conditional ETA (ETA_prim). This is the random effect adjusted for the individual's specific covariate effects.

ETA*: One column for each parameter's raw ETA as reported in the FREM model output.

Covariate Columns: Columns for each covariate, containing the individual covariate estimates. If covmodel = "linear", these values are on the original covariate scale; otherwise, they are on the ETA scale.

Details

The function collects the ETAs from the output of a FREM model, both for the parameters as well as the covariates. The corresponding ETA_prims for the parameter ETAs are computed by extracting the corresponding individual covariate coefficient provided in the FFEMData object.

See also

Examples

# 1. Define the path to the model files included with the package
model_dir <- system.file("extdata/SimNeb/", package = "PMXFrem")

# 2. Load and prepare the dataset also included with the package
data_path <- system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv",
                         package = "PMXFrem")

my_data <- read.csv(data_path)

# In this example dataset, BLQ=1 rows are excluded
my_data <- my_data[my_data$BLQ != 1, ]

# 3. Call calcEtas, providing the data directly
# The function will create the FFEMData object internally.
individual_etas <- calcEtas(
  modName          = "run31",
  modDevDir        = model_dir,
  numNonFREMThetas = 7,
  numSkipOm        = 2,
  dataFile         = my_data,
  parNames         = c("CL", "V", "MAT")
)
#> `FFEMData` object not provided. Creating it internally...

# 4. Display the first few rows of the resulting data frame
# The output contains subject IDs, ETAs, ETA_PRIMs, and covariate values.
head(individual_etas)
#>   ID  ETA1_PRIM    ETA2_PRIM   ETA3_PRIM   ETA4_PRIM   ETA5_PRIM       ETA1
#> 1  1  0.3558900  0.000265641  0.29839265 -0.16738472 -0.45251504  0.3558900
#> 2 13  0.0576561  0.001550910 -0.01011739  0.49976575  0.17628561  0.0576561
#> 3 19  0.1994990  0.000774643 -0.12247343 -0.45317117  0.18602502  0.1994990
#> 4 25  0.0763564  0.000922905  0.35132476 -0.08274008 -0.30447824  0.0763564
#> 5 31 -0.0767972 -0.000963676  0.06676766 -0.05845428  0.14419554 -0.0767972
#> 6 37  0.2059040 -0.000853840 -0.14691612 -0.05886785  0.03829192  0.2059040
#>           ETA2      ETA3      ETA4      ETA5        WT       HT     LBWT
#> 1  0.000265641  0.146266 -0.311718 -0.543485  71.20000 180.6003 58.42542
#> 2  0.001550910  0.536236  0.843135  0.238412  91.20002 175.3004 65.67537
#> 3  0.000774643 -0.190137 -0.743910  0.169696  70.30800 170.0003 55.44496
#> 4  0.000922905  0.310054 -0.327351 -0.319993  64.86500 132.1003 40.48940
#> 5 -0.000963676 -0.476123 -0.582474  0.235579 103.00000 181.6003 72.12320
#> 6 -0.000853840  0.515172  0.525226  0.114228 117.10000 167.6003 66.32499
#>        BSA      AGE      AST      ALT     BILI     CRCL      BMI      SEX
#> 1 1.889940 39.00003 29.62327 25.00005 23.94601  95.3004 21.80003 0.999999
#> 2 2.107348 19.00000 23.99996 25.00004 10.32121 150.0004 29.70003 1.000000
#> 3 1.822091 28.99930 22.99996 17.00000 10.26199 113.7004 24.30003 1.000003
#> 4 1.542820 21.00000 10.00000 12.00010 11.97299 129.4004 37.20003 1.000000
#> 5 2.279418 30.00000 23.99996 48.00000 11.97300 148.7004 31.20003 1.000002
#> 6 2.334910 40.00002 12.90590 12.00000  6.84160 150.0004 41.70000 0.999999
#>     RACEL_3   RACEL_2     NCIL_2    NCIL_1    GENO2    ETHNIC       SMOK
#> 1 -3.00e-07  0.999994 -0.0000041  0.999997 1.999991 0.1033750 -0.0000028
#> 2 -5.00e-07 -0.000002  0.0000071 -0.000002 1.999992 0.3097710  0.9999984
#> 3  1.06e-05  0.999998 -0.0000030  0.000003 1.391606 0.4159864  0.0000012
#> 4  7.70e-06 -0.000001  0.0000109  0.000007 2.065305 1.4194160 -0.0000005
#> 5  3.70e-06  0.000003  0.0296805  0.449148 1.000007 0.4243555 -0.0000033
#> 6  2.00e-06 -0.000003  0.0000019  0.000000 1.999994 0.4211168  1.0000044