Create a new FFEM data set with the FREM covariat effect appended as columns

Add combined FFEM coefficients to the data set for the specified covariates.

Usage

createFFEMdata(
  runno = NULL,
  numNonFREMThetas,
  modName = NULL,
  numFREMThetas = length(grep("THETA", names(dfext))) - numNonFREMThetas,
  covSuffix = "FREMCOV",
  parNames = paste("Par", 1:numParCov, sep = ""),
  numParCov = NULL,
  numSkipOm = 0,
  dataFile,
  newDataFile = paste("vpcData", runno, ".csv", sep = ""),
  filterString = NULL,
  availCov = "all",
  idvar = "ID",
  modDevDir = NULL,
  quiet = FALSE,
  cores = 1,
  dfext = NULL,
  ...
)

Arguments

runno

The run number. Must be provided if modName is NULL. Will be used to create NONMEM model and results file names with the structure run#.mod and run#. NONMEM ext, phi and cov files will be assumed to be named run# followed by the appropriate suffix (where # is runno).

numNonFREMThetas

Number of thetas in the FREM model that are not FREM covariates. These need to come before the FREM covariate thetas in the FREM model file.

modName

The model name. Will be used together with modExt, lstExt, ext, phi and cov to form the NONMEM file names, for example modName.mod for the model file.

numFREMThetas

The number thetas associated with FREM covariate effects.

covSuffix

Use to create the column name fo rthe covariate effects columns. The column names will be the parameter name followed by this string.

parNames

Names of the parameters. Only used in the STDOUT when quiet =FALSE or in eqFile if it is non-blank. Defaults to Par folowed by a number.

numParCov

Number of parameters for which covariate relations are sought (often the same as numNonFREMThetas). default (NULL) indicates that this is guessed by the function by assuming that each covariate effect is modeled as one theta and one eta.

numSkipOm

Number of diagonal omegas that should not be part of the FREM calculations. Such omegas has to come before the large FREM omega block in the FREM model file.

dataFile

The name of the data file used in the base model, i.e. the original data file, or a data.frame with the same data.

newDataFile

The name of a new data file with the FFEM columns added. Default is vpcDatarunno.csv. If NULL, will return a data frame with the data instead of writing it to disk.

filterString

A character string with a filter expression to subset the amended FFEM data set so that the number of rows matches the original data file. Useful if IGNORE statements were used in the original model file.

availCov

Names of the covariates to use in the calculation of the FFEM model. Should be covariate names used in the FREM model.

idvar

The name of the ID column,

modDevDir

The directory where the files reside. Default is the current directory, i.e. ".".

quiet

If FALSE, will print the FFEM model + associated $OMEGA BLOCK to STDOUT.

cores

How many cores to use in the calculations of the FFEM expressions.

dfext

A data frame of the ext file for the FREM model. In case of multiple $EST, the dfext should contain the estimates of the one relevant $EST.

Value

A list with objects:

Omega: The omega prim matrix to be used in the model file. The upper triangle is set to NAs.

Coefficients: An nPar x nCov matrix with covariate coefficients used to compute the individual covariate effects.

indCovEff: A character vector with the labels for the individual covariate effects. Used for the new columns in the new data set.

newData: A data.frame with the new data set, i.e. the same as dataFile with individual covariate effects appended.

Details

This function computes a combined term of all (FREM) covariate effects for each parameter in the FFEM model (individual covariate coefficients) and creates a new data file with these combined effects appended. The individual covariate coefficients are added to the FFEM data set with names that appends COV to the parameter name, e.g. the column with the combined covariate effects for parameter PAR will be called PARCOV. All rows for an individual in the data set will have the same value for PARCOV.

The function calcFFEM() is used to calculate the individual covariate coefficients. It is called once per individual in the data set and takes any missing covariates into account. This means that two subjects with identical covariate values, except that one subject has a missing value for one of the covariate, will have different covariate coefficients for the non-missing covariates.

If the original NONMEM model file use IGNORE statements to subset the data file, it is necessary to provide a filter expression as a character string to obtain an FFEM data set that only include the rows used in the original model. This is done with the filterString argument, e.g. STUDYID==7 to only include STUDYID 7 (this is the same as saying IGNORE.NE.7 in the NM-TRAN model file.) If IGNORE statements are used in original model file and filterString is not specified, then the resulting FFEM data set will have the same number of rows as the the original data set.

Side effetcs

If newDataFile is not NULL, newData will be printed as a csv to disk with the name newDataFile.

If quiet is FALSE, the FFEM expressions and new OMEGA matrix is printed in the console.

See also

Examples

data <- read.csv(system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv", package = "PMXFrem")) %>%
  filter(BLQ != 1)
#> Error in read.csv(system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv",     package = "PMXFrem")) %>% filter(BLQ != 1): could not find function "%>%"

## Check with specified parameter names
ffemData <- createFFEMdata(
  modName          = "run31",
  modDevDir        = system.file("extdata/SimNeb/", package = "PMXFrem"),
  parNames         = c("CL", "V", "MAT"),
  numNonFREMThetas = 7,
  numSkipOm        = 2,
  dataFile         = data,
  newDataFile      = NULL,
  quiet            = TRUE)
#> Error in fread(dataFile, h = TRUE, data.table = FALSE, showProgress = FALSE): input= must be a single character string containing a file name, a system command containing at least one space, a URL starting 'http[s]://', 'ftp[s]://' or 'file://', or, the input data itself containing at least one \n or \r