Generate the input to a parameter table for a FREM model

The data for a parameter estimates table for a FREM model is assembled, including the calculation of the appropriate omega prims given the covariates to take into account. It also computes the covariate coefficients (Individual Covariate Effects) mapping each covariate to each structural parameter.

Usage

fremParameterTable(
  runno = NULL,
  modDevDir = NULL,
  numNonFREMThetas,
  numSkipOm = 0,
  thetaNum,
  omegaNum,
  sigmaNum,
  thetaLabels = paste0("THETA", thetaNum),
  omegaLabels = paste0("OMEGA", omegaNum),
  sigmaLabels = paste0("SIGMA", sigmaNum),
  parNames = NULL,
  covLabels = NULL,
  omegaSD = TRUE,
  sigmaSD = TRUE,
  includeRSE = FALSE,
  includeShrinkage = FALSE,
  shrinkageType = "ETA_SD",
  rawShrinkage = FALSE,
  ffemModName = NULL,
  shkDigs = 2,
  uncertainty = c("RSE", "CI"),
  ciLevel = 0.9,
  sigDigs = 3,
  bsFile = NULL,
  n = 175,
  availCov = "all",
  dfext = NULL,
  modName = NULL,
  modExt = ".mod",
  lstExt = ".lst",
  quiet = FALSE,
  seed = NULL,
  ...
)

Arguments

runno

The run number. Must be provided if modName is NULL. Will be used to create NONMEM model and results file names with the structure run#.mod and run#. NONMEM ext, phi and cov files will be assumed to be named run# followed by the appropriate suffix (where # is runno).

modDevDir

The directory where the files reside. Default is the current directory, i.e. ".".

numNonFREMThetas

Number of thetas in the FREM model that are not FREM covariates. These need to come before the FREM covariate thetas in the FREM model file.

numSkipOm

Number of diagonal omegas that should not be part of the FREM calculations. Such omegas has to come before the large FREM omega block in the FREM model file.

thetaNum

The theta numbers to include in the output. This refers to the number associated with the theta value in parVector. `

omegaNum

The omega numbers to include in the output. This refers to the number associated with the omega value in parVector. `

sigmaNum

The sigma numbers to include in the output. This refers to the number associated with the omega value in parVector. `

thetaLabels

A character vector with labels for the THETAs. Should be exactly as long as thetaNum. Note: This is used strictly for the base parameter table, which may include non-FREM structural parameters.

omegaLabels

A vector with labels for the OMEGAs. Should be as long as omegaNum.

sigmaLabels

A vector with labels for the SIGMAs Should be as long as sigmaNum.

parNames

A character vector with labels for the structural parameters affected by covariates (which will become the rows in the coefficient table). Should be exactly as long as numParCov. If NULL, defaults to Par1, Par2, etc. Note: Do not confuse this with thetaLabels, as the number of structural ETAs (numParCov) often differs from the total thetaNum.

covLabels

A character vector with labels for the covariates (columns in the coefficient table). Should match the length of availCov.

omegaSD

Logical. If the omega prims should be reported on the SD scale (i.e. the square root of omega prim). Default is TRUE.

sigmaSD

Logical. If the sigmas should be reported on the SD scale (i.e. the square root of sigma). Default is TRUE.

includeRSE

Logical. Should the output include relative standard errors (RSE). Default is FALSE.

includeShrinkage

Logical. If TRUE, calculates and includes shrinkages in the table using a specified FFEM model output. Defaults to FALSE.

shrinkageType

Character string specifying which type of shrinkage to report. Options are "ETA_Var", "ETA_SD", "EBV_Var", or "EBV_SD". Defaults to "ETA_SD" (the legacy NONMEM ETA SD shrinkage).

rawShrinkage

Logical. If FALSE (default), negative shrinkages are floored and reported as 1.0000e-10 to exactly mimic NONMEM's standard output. If TRUE, the raw (potentially negative) calculated shrinkages are reported.

ffemModName

A character string specifying the model name of the FFEM run (typically a MAX=0 run) to be used exclusively for shrinkage calculations. Required if includeShrinkage is TRUE.

shkDigs

An integer specifying the number of decimal places to use when formatting the shrinkage output. Default is 2.

uncertainty

A character string specifying how uncertainty should be summarized. Either "RSE" (default) or "CI".

ciLevel

A numeric value between 0 and 1 specifying the confidence interval level. Default is 0.90.

sigDigs

An integer specifying the number of significant digits to use for formatting output estimates and uncertainties. Default is 3.

bsFile

The name of a PsNbootstrap or sir file raw_results file. To be used for RSE calculations based on bootstrap output.

n

The number of samples to use in the RSE calculations.

availCov

Names of the covariates to use in the calculation of the FFEM model. Should be covariate names used in the FREM model.

dfext

A data frame of the ext file for the FREM model. In case of multiple $EST, the dfext should contain the estimates of the one relevant $EST.

modName

The model name. Will be used together with modExt, lstExt, ext, phi and cov to form the NONMEM file names, for example modName.mod for the model file.

modExt

The model file name extension. Default is ".mod".

lstExt

The NONMEM out file name extension. Default is ".lst".

quiet

If FALSE, will print the FFEM model + associated $OMEGA BLOCK to STDOUT.

...

Additional arguments passed directly to PMXForest::getSamples() (e.g., seed for reproducibility).

Value

A list of five components:

parameterTable: A data frame with information for the base parameter table.

coefficientTable_long: A tidy, long-format data frame of the covariate coefficients (optimal for internal reporting APIs).

coefficientTable_wide: A document-ready wide-format data frame of the covariate coefficients (Covariates x Parameters). Estimates and uncertainties are separated into distinct columns.

Samples: A data frame with the parameter vector samples used to derive the RSE information.

Condition: The condition number for the variance-covariance matrix of the parameters.

See also

Examples

if (requireNamespace("kableExtra", quietly = TRUE)) {
library(kableExtra)
library(dplyr)

set.seed(123)
runno            <- 31
modName          <- "run31"
ffemModName      <- "run31max0"
modDevDir        <- system.file("extdata/SimNeb/",package = "PMXFrem")
numNonFREMThetas <- 7
numSkipOm        <- 2

## Generate the raw output
tmp <- fremParameterTable(
      runno            = runno,
      modName          = modName,
      modDevDir        = modDevDir,
      thetaNum         = 2:7,
      omegaNum         = c(1,3,4,5),
      sigmaNum         = 1,
      thetaLabels      = c("CL (L/h)","V (L)","MAT (h)","D1 (h)","Food on Frel","Food on MAT"),
      omegaLabels      = c("IIV on RUV","IIV on CL","IIV on V","IIV on MAT"),
      sigmaLabels      = c("RUV"),
      parNames         = c("CL", "V", "MAT"),
      includeRSE       = TRUE,
      includeShrinkage = TRUE,
      shrinkageType    = "ETA_SD",
      rawShrinkage     = FALSE,
      ffemModName      = ffemModName,
      shkDigs          = 2,
      uncertainty      = "CI",
      numNonFREMThetas = numNonFREMThetas,
      numSkipOm        = numSkipOm,
      availCov         = "all",
      quiet            = TRUE)

## Use kable_extra to generate a nice looking table of the main parameters
tmp$parameterTable %>%
    mutate(Estimate=as.character(signif(Estimate,3)))%>%
    kbl() %>%
    kable_classic(full_width = FALSE, html_font = "Cambria")
    
## Use kable_extra to generate a nice looking table of the covariate coefficients
tmp$coefficientTable_wide %>%
    kbl() %>%
    kable_classic(full_width = FALSE, html_font = "Cambria")
}
#> 
#> Attaching package: ‘kableExtra’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     group_rows
#> <table class=" lightable-classic" style="font-family: Cambria; width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Covariate </th>
#>    <th style="text-align:left;"> CL </th>
#>    <th style="text-align:left;"> CL 90% CI </th>
#>    <th style="text-align:left;"> V </th>
#>    <th style="text-align:left;"> V 90% CI </th>
#>    <th style="text-align:left;"> MAT </th>
#>    <th style="text-align:left;"> MAT 90% CI </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> WT </td>
#>    <td style="text-align:left;"> -0.00591 </td>
#>    <td style="text-align:left;"> [-0.0334 - 0.0166] </td>
#>    <td style="text-align:left;"> -0.0265 </td>
#>    <td style="text-align:left;"> [-0.0567 - -0.00872] </td>
#>    <td style="text-align:left;"> 0.00556 </td>
#>    <td style="text-align:left;"> [-0.0239 - 0.0326] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> HT </td>
#>    <td style="text-align:left;"> -0.0287 </td>
#>    <td style="text-align:left;"> [-0.0566 - 0.00653] </td>
#>    <td style="text-align:left;"> -0.0279 </td>
#>    <td style="text-align:left;"> [-0.0701 - -0.00379] </td>
#>    <td style="text-align:left;"> -0.0102 </td>
#>    <td style="text-align:left;"> [-0.0596 - 0.0303] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> LBWT </td>
#>    <td style="text-align:left;"> -0.0181 </td>
#>    <td style="text-align:left;"> [-0.0434 - 0.00489] </td>
#>    <td style="text-align:left;"> -0.0111 </td>
#>    <td style="text-align:left;"> [-0.0535 - 0.0133] </td>
#>    <td style="text-align:left;"> -0.000679 </td>
#>    <td style="text-align:left;"> [-0.0295 - 0.0327] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BSA </td>
#>    <td style="text-align:left;"> 2.75 </td>
#>    <td style="text-align:left;"> [-0.684 - 6.36] </td>
#>    <td style="text-align:left;"> 4.01 </td>
#>    <td style="text-align:left;"> [0.844 - 9.18] </td>
#>    <td style="text-align:left;"> 0.202 </td>
#>    <td style="text-align:left;"> [-4.75 - 5.50] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AGE </td>
#>    <td style="text-align:left;"> 0.00215 </td>
#>    <td style="text-align:left;"> [-0.00128 - 0.00551] </td>
#>    <td style="text-align:left;"> -0.000617 </td>
#>    <td style="text-align:left;"> [-0.00466 - 0.00296] </td>
#>    <td style="text-align:left;"> 0.00319 </td>
#>    <td style="text-align:left;"> [-0.000881 - 0.00796] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AST </td>
#>    <td style="text-align:left;"> -0.00466 </td>
#>    <td style="text-align:left;"> [-0.0119 - -0.000369] </td>
#>    <td style="text-align:left;"> 0.000291 </td>
#>    <td style="text-align:left;"> [-0.00512 - 0.00509] </td>
#>    <td style="text-align:left;"> -0.00894 </td>
#>    <td style="text-align:left;"> [-0.0186 - -0.00237] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> ALT </td>
#>    <td style="text-align:left;"> 0.00427 </td>
#>    <td style="text-align:left;"> [0.00101 - 0.00896] </td>
#>    <td style="text-align:left;"> 0.000161 </td>
#>    <td style="text-align:left;"> [-0.00319 - 0.00452] </td>
#>    <td style="text-align:left;"> 0.00616 </td>
#>    <td style="text-align:left;"> [0.00232 - 0.0129] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BILI </td>
#>    <td style="text-align:left;"> 0.00201 </td>
#>    <td style="text-align:left;"> [-0.00615 - 0.0112] </td>
#>    <td style="text-align:left;"> 0.000162 </td>
#>    <td style="text-align:left;"> [-0.00904 - 0.00895] </td>
#>    <td style="text-align:left;"> 0.00228 </td>
#>    <td style="text-align:left;"> [-0.00683 - 0.0133] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> CRCL </td>
#>    <td style="text-align:left;"> 0.00117 </td>
#>    <td style="text-align:left;"> [-0.000874 - 0.00310] </td>
#>    <td style="text-align:left;"> -0.000471 </td>
#>    <td style="text-align:left;"> [-0.00279 - 0.00154] </td>
#>    <td style="text-align:left;"> 0.00189 </td>
#>    <td style="text-align:left;"> [-0.000470 - 0.00418] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BMI </td>
#>    <td style="text-align:left;"> -0.0500 </td>
#>    <td style="text-align:left;"> [-0.108 - 0.0208] </td>
#>    <td style="text-align:left;"> -0.0146 </td>
#>    <td style="text-align:left;"> [-0.0963 - 0.0362] </td>
#>    <td style="text-align:left;"> -0.0274 </td>
#>    <td style="text-align:left;"> [-0.116 - 0.0577] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> SEX </td>
#>    <td style="text-align:left;"> -0.168 </td>
#>    <td style="text-align:left;"> [-0.367 - 0.0507] </td>
#>    <td style="text-align:left;"> 0.153 </td>
#>    <td style="text-align:left;"> [-0.145 - 0.342] </td>
#>    <td style="text-align:left;"> -0.0419 </td>
#>    <td style="text-align:left;"> [-0.247 - 0.248] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> RACEL_3 </td>
#>    <td style="text-align:left;"> 0.00281 </td>
#>    <td style="text-align:left;"> [-0.108 - 0.142] </td>
#>    <td style="text-align:left;"> 0.0197 </td>
#>    <td style="text-align:left;"> [-0.132 - 0.154] </td>
#>    <td style="text-align:left;"> 0.0160 </td>
#>    <td style="text-align:left;"> [-0.175 - 0.199] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> RACEL_2 </td>
#>    <td style="text-align:left;"> 0.00800 </td>
#>    <td style="text-align:left;"> [-0.0602 - 0.0752] </td>
#>    <td style="text-align:left;"> 0.00167 </td>
#>    <td style="text-align:left;"> [-0.0815 - 0.0727] </td>
#>    <td style="text-align:left;"> 0.0392 </td>
#>    <td style="text-align:left;"> [-0.0300 - 0.127] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> NCIL_2 </td>
#>    <td style="text-align:left;"> -0.517 </td>
#>    <td style="text-align:left;"> [-0.788 - -0.291] </td>
#>    <td style="text-align:left;"> -0.0215 </td>
#>    <td style="text-align:left;"> [-0.414 - 0.235] </td>
#>    <td style="text-align:left;"> -0.174 </td>
#>    <td style="text-align:left;"> [-0.421 - 0.119] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> NCIL_1 </td>
#>    <td style="text-align:left;"> -0.238 </td>
#>    <td style="text-align:left;"> [-0.349 - -0.139] </td>
#>    <td style="text-align:left;"> 0.0227 </td>
#>    <td style="text-align:left;"> [-0.0962 - 0.116] </td>
#>    <td style="text-align:left;"> -0.106 </td>
#>    <td style="text-align:left;"> [-0.238 - 0.00490] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> GENO2 </td>
#>    <td style="text-align:left;"> 0.737 </td>
#>    <td style="text-align:left;"> [0.690 - 0.792] </td>
#>    <td style="text-align:left;"> 0.734 </td>
#>    <td style="text-align:left;"> [0.686 - 0.792] </td>
#>    <td style="text-align:left;"> 0.0128 </td>
#>    <td style="text-align:left;"> [-0.0575 - 0.0712] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> ETHNIC </td>
#>    <td style="text-align:left;"> -0.0533 </td>
#>    <td style="text-align:left;"> [-0.106 - 6.12e-05] </td>
#>    <td style="text-align:left;"> -0.0631 </td>
#>    <td style="text-align:left;"> [-0.123 - 0.00600] </td>
#>    <td style="text-align:left;"> -0.0158 </td>
#>    <td style="text-align:left;"> [-0.0773 - 0.0501] </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> SMOK </td>
#>    <td style="text-align:left;"> 0.367 </td>
#>    <td style="text-align:left;"> [0.265 - 0.466] </td>
#>    <td style="text-align:left;"> 0.250 </td>
#>    <td style="text-align:left;"> [0.116 - 0.401] </td>
#>    <td style="text-align:left;"> 0.0664 </td>
#>    <td style="text-align:left;"> [-0.0839 - 0.218] </td>
#>   </tr>
#> </tbody>
#> </table>