Collect filenames that are associated with a NONMEM run

Assembles the model, ext, phi and lst file names for a run based on either run number or model name (the part of the file name to the left of the filename extension).

Usage

getFileNames(
  runno = NULL,
  modName = NULL,
  modDevDir = ".",
  modExt = ".mod",
  lstExt = ".lst"
)

Arguments

runno

The run number. Must be provided if modName is NULL. Will be used to create NONMEM model and results file names with the structure run#.mod and run#. NONMEM ext, phi and cov files will be assumed to be named run# followed by the appropriate suffix (where # is runno).

modName

The model name. Will be used together with modExt, lstExt, ext, phi and cov to form the NONMEM file names, for example modName.mod for the model file.

modDevDir

The directory where the files reside. Default is the current directory, i.e. ".".

modExt

The model file name extension. Default is ".mod".

lstExt

The NONMEM out file name extension. Default is ".lst".

Value

A list with components "mod", "ext", "phi" and "lst" specifying the path names of the corresponding files.

Details

runno and modName can not both be NULL. If both are provided, modName is used.

modDevDir can not be NULL.

See also

Examples

getFileNames(runno = 30, modDevDir = system.file("extdata/SimNeb/", package = "PMXFrem"))
#> $mod
#> [1] "/home/github/runner/_work/_temp/Rlib-32344486548/PMXFrem/extdata/SimNeb//run30.mod"
#> 
#> $ext
#> [1] "/home/github/runner/_work/_temp/Rlib-32344486548/PMXFrem/extdata/SimNeb//run30.ext"
#> 
#> $phi
#> [1] "/home/github/runner/_work/_temp/Rlib-32344486548/PMXFrem/extdata/SimNeb//run30.phi"
#> 
#> $lst
#> [1] "/home/github/runner/_work/_temp/Rlib-32344486548/PMXFrem/extdata/SimNeb//run30.lst"
#> 
#> $cov
#> [1] "/home/github/runner/_work/_temp/Rlib-32344486548/PMXFrem/extdata/SimNeb//run30.cov"
#>