Modify an exiting FREM model and data set

Add or remove covariates and/or data to a FREM model and data set.

Usage

updateFREMmodel(
  strFREMModel,
  strFREMData,
  strFFEMData,
  cstrContCovsToAdd = NULL,
  cstrCatCovsToAdd = NULL,
  cstrCovsToAddOrder = NULL,
  strNewFREMData = NULL,
  filterString = NULL,
  strUpdateType = "DataAndModel",
  quiet = TRUE,
  strID = "ID",
  basenames_th = NULL,
  basenames_om = NULL,
  numNonFREMThetas,
  numSkipOm = 0,
  numParCov = NULL,
  cstrKeepCols = c("ID", "TIME", "AMT", "DV", "II", "EVID", "SS", "RATE", "FREMTYPE"),
  cstrSetToZero = c("AMT", "II", "SS", "EVID", "RATE"),
  bWriteData = TRUE,
  bWriteFIX = TRUE,
  bWriteMod = TRUE,
  cstrDV = "DV",
  cstrRemoveCov = NULL,
  covEpsNum = 2,
  overrideExistingCheck = FALSE,
  bRecodeDichotomous = FALSE,
  allowNon01 = FALSE,
  keepDoseOnlySubjects = FALSE,
  sortFREMDataset = NULL
)

Arguments

strFREMModel

File name of the FREM-model to add/remove covariates to/from.

strFREMData

Name of FREM-dataset (file or data.frame) to add/remove covariates to/from (not used with strUpdateType "NoData").

strFFEMData

Name of FFEM-dataset (normal dataset, (file or data.frame)) that will be used to append the FREM-dataset, (not used with strUpdateType "NoData").

cstrContCovsToAdd

A vector of continuous covariate names to add, default = NULL, (not used with strUpdateType "NoData").

cstrCatCovsToAdd

A vector of categorical covariate names to add, default = NULL, (not used with strUpdateType "NoData").

cstrCovsToAddOrder

A vector of the order of the covariate names to add, default = NULL (i.e. alphabetic order will be used), (not used with strUpdateType "NoData"). Note: if used, this should contain all covariates in cstrCatCovsToAdd as well as cstrContCovsToAdd.

strNewFREMData

Name of the new dataset, default=paste0(strFREMData_without_extension,"new",".",extension), (not used with strUpdateType "NoData").

strUpdateType

Update function to run: "DataAndModel" - Create new data and add/remove variables from the model (with updated inits). "NoData" - Do not create data or add variables to model, only update the frem model in terms of new inits

quiet

If set to FALSE, the function outputs verbose information on what it is doing.

strID

A string with the ID identifier column in the FFEM dataset.

basenames_th

A vector of strings with the names of the base variables (used for commenting thetas), should be the same length as number of nonFREMThetas in the model, if NULL, BASE1,BASE2,etc are used as names.

basenames_om

A vector of strings with the names of the base variables (used for commenting omegas), should be the same length as number of numSkipOm+numParCov in the model, if NULL, BASE1,BASE2,etc are used as names.

numNonFREMThetas

Number of thetas in the FREM model that are not FREM covariates. These need to come before the FREM covariate thetas in the FREM model file.

numSkipOm

Number of diagonal omegas that should not be part of the FREM calculations. Such omegas has to come before the large FREM omega block in the FREM model file.

numParCov

Number of parameters for which covariate relations are sought (often the same as numNonFREMThetas). default (NULL) indicates that this is guessed by the function by assuming that each covariate effect is modeled as one theta and one eta.

cstrKeepCols

A vector of columns to keep in the dataset (for the updated new dataset).

cstrSetToZero

A vector of variable names that should be set to zero on the covariate FREMTYPE lines. Typically columns with PK dosing information.

bWriteData

If FALSE; add new variables to the model file but do not write new datasets, has no effect when "NoData" is used.

bWriteFIX

If TRUE; FIX is written to the theta parameter estimates code for the covariates that were fixed in the model file, if FALSE; all theta parameters are assumed to be estimated.

bWriteMod

If TRUE; write the new model file to disk with _mod appended to the file name (before the suffix).

cstrDV

A vector of strings with DV variables that should be added, default="DV" (i.e. DV assumed fremtype=0), additional DVs are added with fremtype 1,2,3...etc.

cstrRemoveCov

A vector of strings for covariates that should be remove, note that FREMTYPEs for remaining covariates might/will change. The removal of covariates are done before any adding of data and/or new covariates. Note that if this functionality is used to remove the last existing category of a categorical covariate, this should be done by removing the orginal name of the covariate and not the specific categorical covariate, i.e. "SITEID" instead of "SITEID_1" to ensure consistent renumbering of FREMTYPEs

covEpsNum

The number of the epsilons parameter to be used for the covariates.

overrideExistingCheck

If TRUE, the existing check will be overriden and covariates will be added even though they are present in $DATA of the modefile

bRecodeDichotomous

Logical. Should dichotomous covariates be automatically recoded to 0/1? Defaults to FALSE. If FALSE, inputs must be strictly 0/1.

allowNon01

Logical. If TRUE, allows non-0/1 dichotomous covariates (like 1/2) to pass through untouched for PsN compatibility. Defaults to FALSE.

keepDoseOnlySubjects

Logical. If FALSE (default), subjects without any valid PK observations (e.g., only dosing records) are completely excluded from the generated dataset. If TRUE, these subjects are retained and their covariates are included as observations.

Value

An invisible list with components data and model, containing the new data set (if any, else NULL) and updated model.

Side effects

Will write the new fremdata set (if bWriteData is TRUE and strUpdateType is not 'NoData') and updated model file (if bwriteMod is TRUE) to disc. The model file name will be 'stem'_new.mod).

See also

Examples

## Example 1: Remove Covariates from a Model
# This example removes "SEX" from the model and data set.
td <- tempdir() # Create a temporary directory for output files

updateFREMmodel(
  strFREMModel      = system.file("extdata/SimNeb/run31.mod", package = "PMXFrem"),
  strFREMData       = system.file("extdata/SimNeb/frem_dataset.dta", package = "PMXFrem"),
  strFFEMData       = system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv", package = "PMXFrem"),
  cstrRemoveCov     = c("SEX"),
  strNewFREMData    = file.path(td, "frem_dataset_noSEX.csv"),
  numNonFREMThetas  = 7,
  numSkipOm         = 2,
  bWriteData        = TRUE,
  bWriteMod         = TRUE,
  quiet             = FALSE,
  bWriteFIX         = TRUE,
  cstrKeepCols      = c("ID", "TIME", "AMT", "EVID", "RATE", "DV", "FOOD", "FREMTYPE")
)
#> Warning: A single-character IGNORE statement was found in the $DATA record.
#> If this filter is used for anything else than managing the header line, you will have to pre-filter the data or use IGNORE=(list).
#> [1] "Found a continuous covariate to remove SEX with FREMTYPE=1100"
#> [1] "Removing all unwanted fremtypes from FREM dataset, in total 730 observations"
#> [1] "Recoding all remaining FREMTYPEs in FREM dataset..."
#> [1] "Updating 691 rows with FREMTYPE=1200 (RACEL_3) to FREMTYPE=1100"
#> [1] "Updating 691 rows with FREMTYPE=1300 (RACEL_2) to FREMTYPE=1200"
#> [1] "Updating 697 rows with FREMTYPE=1400 (NCIL_2) to FREMTYPE=1300"
#> [1] "Updating 697 rows with FREMTYPE=1500 (NCIL_1) to FREMTYPE=1400"
#> [1] "Updating 475 rows with FREMTYPE=1600 (GENO2) to FREMTYPE=1500"
#> [1] "Updating 665 rows with FREMTYPE=1700 (ETHNIC) to FREMTYPE=1600"
#> [1] "Updating 730 rows with FREMTYPE=1800 (SMOK) to FREMTYPE=1700"
#> [1] "Recoding FREMTYPES, done!"
#> [1] "Variables already in FREM model (n=15): AGE ALT AST BILI BMI BSA CRCL ETHNIC GENO2 HT LBWT NCIL RACEL SMOK WT"
#> [1] "Found 0 individuals that should be added to the FREM dataset"
#> [1] "No observations for DV (fremtype=0); not adding any observations!"
#> Data integrity validation passed: The FREM dataset perfectly mirrors the base dataset.

## Example 2: Add Covariates to a Model
# This example adds a new categorical covariate, "SITE", to the model.
td <- tempdir()

# First, create a temporary FFEM dataset that includes the new covariate
ffem_df <- as.data.frame(data.table::fread(
  system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv", package = "PMXFrem")
))
ffem_df$SITE <- rep(c(101, 102, 103), length.out = nrow(ffem_df)) # Add 3-level SITE covariate

# Define paths to original model and FREM data
model_path <- system.file("extdata/SimNeb/run31.mod", package = "PMXFrem")
frem_data_path <- system.file("extdata/SimNeb/frem_dataset.dta", package = "PMXFrem")

# Copy the .ext file to the temp directory so the function can find it
file.copy(system.file("extdata/SimNeb/run31.ext", package = "PMXFrem"), td)
#> [1] TRUE
file.copy(model_path, td)
#> [1] TRUE
model_in_td <- file.path(td, "run31.mod")

updateFREMmodel(
  strFREMModel      = model_in_td,
  strFREMData       = frem_data_path,
  strFFEMData       = ffem_df, # Use the modified data frame with SITE
  strNewFREMData    = file.path(td, "frem_data_with_SITE.csv"),
  cstrCatCovsToAdd  = "SITE",  # Add the new SITE covariate
  numNonFREMThetas  = 7,
  numSkipOm         = 2,
  bWriteData        = TRUE,
  bWriteMod         = TRUE,
  quiet             = FALSE
)
#> Warning: A single-character IGNORE statement was found in the $DATA record.
#> If this filter is used for anything else than managing the header line, you will have to pre-filter the data or use IGNORE=(list).
#> [1] "Variables already in FREM model (n=16): AGE ALT AST BILI BMI BSA CRCL ETHNIC GENO2 HT LBWT NCIL RACEL SEX SMOK WT"
#> [1] "Categorical covariates that will be added to FREM (n=1): SITE"
#> [1] "Identifying new covariate: SITE_102 with fremtype 1900"
#> [1] "Identifying new covariate: SITE_103 with fremtype 2000"
#> [1] "Found 0 individuals that should be added to the FREM dataset"
#> [1] "No observations for DV (fremtype=0); not adding any observations!"
#> [1] "Adding 732 covariate values (SITE_102) from 732 individuals as fremtype 1900"
#> [1] "Adding 732 covariate values (SITE_103) from 732 individuals as fremtype 2000"
#> Data integrity validation passed: The FREM dataset perfectly mirrors the base dataset.

## Example 3: Add New Individuals to an Existing FREM Dataset
td <- tempdir()

# Load the original data as data frames
frem_df <- as.data.frame(data.table::fread(
  system.file("extdata/SimNeb/frem_dataset.dta", package = "PMXFrem")
)) %>% rename(LNDV=ODV)
ffem_df <- as.data.frame(data.table::fread(
  system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv", package = "PMXFrem")
)) 

# Simulate new individuals by taking 50 rows and giving them new IDs
ffem_new_ids <- ffem_df[1:50, ]
ffem_new_ids$ID <- ffem_new_ids$ID + max(frem_df$ID)

# Use the original model
model_path <- system.file("extdata/SimNeb/run31.mod", package = "PMXFrem")

result <- updateFREMmodel(
  strFREMModel      = model_path,
  strFREMData       = frem_df,      # Pass original data as a data frame
  strFFEMData       = ffem_new_ids, # Pass the new individuals' data
  strNewFREMData    = file.path(td, "frem_data_with_new_ids.csv"),
  numNonFREMThetas  = 7,
  numSkipOm         = 2,
  bWriteData        = TRUE,
  bWriteMod         = FALSE, # Model structure doesn't change
  quiet             = FALSE
)
#> Warning: A single-character IGNORE statement was found in the $DATA record.
#> If this filter is used for anything else than managing the header line, you will have to pre-filter the data or use IGNORE=(list).
#> [1] "Variables already in FREM model (n=16): AGE ALT AST BILI BMI BSA CRCL ETHNIC GENO2 HT LBWT NCIL RACEL SEX SMOK WT"
#> [1] "Found 4 individuals that should be added to the FREM dataset"
#> [1] "Adding 50 observations (DV) from 4 individuals as fremtype 0"
#> [1] "Adding 4 continuous covariate values (WT) from 4 individuals as fremtype 100"
#> [1] "Adding 4 continuous covariate values (HT) from 4 individuals as fremtype 200"
#> [1] "Adding 4 continuous covariate values (LBWT) from 4 individuals as fremtype 300"
#> [1] "Adding 4 continuous covariate values (BSA) from 4 individuals as fremtype 400"
#> [1] "Adding 3 continuous covariate values (AGE) from 3 individuals as fremtype 500"
#> [1] "Adding 3 continuous covariate values (AST) from 3 individuals as fremtype 600"
#> [1] "Adding 4 continuous covariate values (ALT) from 4 individuals as fremtype 700"
#> [1] "Adding 3 continuous covariate values (BILI) from 3 individuals as fremtype 800"
#> [1] "Adding 4 continuous covariate values (CRCL) from 4 individuals as fremtype 900"
#> [1] "Adding 4 continuous covariate values (BMI) from 4 individuals as fremtype 1000"
#> [1] "Adding 4 continuous covariate values (SEX) from 4 individuals as fremtype 1100"
#> [1] "Adding 4 categorical covariate values (RACEL_3) from 4 individuals as fremtype 1200"
#> [1] "Adding 4 categorical covariate values (RACEL_2) from 4 individuals as fremtype 1300"
#> [1] "Adding 4 categorical covariate values (NCIL_2) from 4 individuals as fremtype 1400"
#> [1] "Adding 4 categorical covariate values (NCIL_1) from 4 individuals as fremtype 1500"
#> [1] "Adding 2 continuous covariate values (GENO2) from 2 individuals as fremtype 1600"
#> [1] "No observed covariate values for ETHNIC (fremtype=1700); not adding any covariate values!"
#> [1] "Adding 4 continuous covariate values (SMOK) from 4 individuals as fremtype 1800"
#> Data integrity validation passed: The FREM dataset perfectly mirrors the base dataset.

# The returned data frame now contains the new IDs
print(tail(result$data))
#>        ID TIME AMT     DV EVID RATE FREMTYPE
#> 45541 989    6   0 49.560    0    0        0
#> 45542 989    8   0 25.143    0    0        0
#> 45543 989   12   0 21.963    0    0        0
#> 45544 989   16   0 16.067    0    0        0
#> 45545 989   24   0  7.048    0    0        0
#> 45546 989   36   0  1.901    0    0        0

## Example 4: Only Update Initial Estimates from an .ext file
td <- tempdir()
model_path <- system.file("extdata/SimNeb/run31.mod", package = "PMXFrem")

# Copy model and ext file to a temporary directory
file.copy(model_path, td)
#> [1] FALSE
file.copy(system.file("extdata/SimNeb/run31.ext", package = "PMXFrem"), td)
#> [1] FALSE
model_in_td <- file.path(td, "run31.mod")

updateFREMmodel(
  strFREMModel       = model_in_td,
  strUpdateType      = "NoData",
  basenames_th       = c("CL","V","MAT","D1","FRELFOOD","MATFOOD"),
  basenames_om       = c("RUV","D1","CL","V","MAT"),
  numNonFREMThetas   = 7,
  numSkipOm          = 2,
  bWriteData         = FALSE,
  bWriteMod          = TRUE,
  bWriteFIX          = TRUE,
  bRecodeDichotomous = TRUE,
  quiet              = FALSE
)