Calculate conditional ETAs (ETA prim)

Collects the ETAs (parameter and covariate) from a FREM model and computes the ETA_prims. Optionally appends the true Empirical Bayes Estimates (EBEs) from a specified FFEM run for diagnostic plotting.

Usage

calcEtas(
  runno = NULL,
  numNonFREMThetas,
  modName = NULL,
  numSkipOm = 0,
  idvar = "ID",
  modDevDir = NULL,
  FFEMData = NULL,
  covmodel = "linear",
  dataFile = NULL,
  parNames = NULL,
  quiet = TRUE,
  ffemModName = NULL,
  appendMissingFlags = TRUE,
  ...
)

Arguments

runno

The run number. Must be provided if modName is NULL. Will be used to create NONMEM model and results file names with the structure run#.mod and run#. NONMEM ext, phi and cov files will be assumed to be named run# followed by the appropriate suffix (where # is runno).

numNonFREMThetas

Number of thetas in the FREM model that are not FREM covariates. These need to come before the FREM covariate thetas in the FREM model file.

modName

The model name. Will be used together with modExt, lstExt, ext, phi and cov to form the NONMEM file names, for example modName.mod for the model file.

numSkipOm

Number of diagonal omegas that should not be part of the FREM calculations. Such omegas has to come before the large FREM omega block in the FREM model file.

idvar

The name of the ID column,

modDevDir

The directory where the files reside. Default is the current directory, i.e. ".".

FFEMData

An FFEMData object as obtained with the function createFFEMdata. If NULL, the object will be created internally, requiring dataFile, parNames, and other arguments for createFFEMdata to be provided.

covmodel

A character string indicating if the covariate models were implemented linearly (additatively) in the frem model or not. Default is "linear".

dataFile

The name of the data file used in the base model, i.e. the original data file, or a data.frame with the same data.

parNames

Names of the parameters. Only used in the STDOUT when quiet =FALSE or in eqFile if it is non-blank. Defaults to Par folowed by a number.

quiet

If FALSE, will print the FFEM model + associated $OMEGA BLOCK to STDOUT.

ffemModName

A character string specifying the model name of the FFEM run (typically a MAXEVAL=0 run). If provided, the true posterior EBEs will be extracted from its .phi file and appended to the output data frame. Default is NULL.

appendMissingFlags

Logical. If TRUE, evaluates the original dataset for missing covariate values (coded as -99 or NA) and appends a binary indicator column for each predicted covariate with the suffix _MISSING (1 = Missing, 0 = Not Missing). Default is FALSE.

...

Additional arguments passed on to createFFEMdata when FFEMData is NULL.

Value

A data.frame with one row per subject, containing the individual parameter estimates. The data frame includes the following columns:

ID: The subject identifier.

ETA_PRIM*: One column for each parameter's conditional ETA (ETA_prim). This is the random effect adjusted for the individual's specific covariate effects.

ETA*: One column for each parameter's raw ETA as reported in the FREM model output.

Covariate Columns: Columns for each covariate, containing the individual covariate estimates. If covmodel = "linear", these values are on the original covariate scale; otherwise, they are on the ETA scale.

EBE_ETA*: (If ffemModName is provided) The true Empirical Bayes Estimate from the FFEM model.

*_MISSING: (If appendMissingFlags is TRUE) A binary indicator of missingness in the original dataset.

Details

The function collects the ETAs from the output of a FREM model, both for the parameters as well as the covariates. The corresponding ETA_prims for the parameter ETAs are computed by extracting the corresponding individual covariate coefficient provided in the FFEMData object. If ffemModName is specified, the true structural EBEs are extracted from the FFEM model and merged by ID.

See also

Examples

# 1. Define the path to the model files included with the package
model_dir <- system.file("extdata/SimNeb/", package = "PMXFrem")

# 2. Load and prepare the dataset also included with the package
data_path <- system.file("extdata/SimNeb/DAT-2-MI-PMX-2-onlyTYPE2-new.csv",
                         package = "PMXFrem")

my_data <- read.csv(data_path)

# In this example dataset, BLQ=1 rows are excluded
my_data <- my_data[my_data$BLQ != 1, ]

# 3. Call calcEtas, providing the data directly
# The function will create the FFEMData object internally.
# We also specify ffemModName to extract true posterior EBEs for plotting.
individual_etas <- calcEtas(
  modName            = "run31",
  modDevDir          = model_dir,
  numNonFREMThetas   = 7,
  numSkipOm          = 2,
  dataFile           = my_data,
  parNames           = c("CL", "V", "MAT"),
  ffemModName        = "run31max0",
  appendMissingFlags = TRUE
)

# 4. Display the first few rows of the resulting data frame
# The output contains subject IDs, ETAs, ETA_PRIMs, EBEs, and missingness flags.
head(individual_etas)
#>   ID ETA1_PRIM    ETA2_PRIM   ETA3_PRIM   ETA4_PRIM  ETA5_PRIM      ETA1
#> 1  1  0.355890  0.000265641  0.29839265 -0.16738472 -0.4525150  0.355890
#> 2  2  0.172514  0.000223009 -0.09115155 -0.01035833  0.2333414  0.172514
#> 3  3  0.197031 -0.001197820 -0.03890302 -0.12012065 -0.1972245  0.197031
#> 4  4 -0.318559  0.000221808 -0.26696247  0.21725668  0.2852994 -0.318559
#> 5  5  0.128488  0.001241110 -0.24828690 -0.13152842 -0.1702444  0.128488
#> 6  6 -0.244566  0.000729417  0.19374638  0.03682627 -0.4478100 -0.244566
#>           ETA2       ETA3       ETA4      ETA5       WT       HT     LBWT
#> 1  0.000265641  0.1462660 -0.3117180 -0.543485 71.20000 180.6003 58.42542
#> 2  0.000223009 -0.2453740 -0.5942760  0.215847 56.00000 172.7003 48.14133
#> 3 -0.001197820  0.0360389  0.0334388 -0.220241 73.60000 168.0003 50.34672
#> 4  0.000221808 -0.7553180  0.0554221  0.108853 83.70003 170.2003 61.11420
#> 5  0.001241110 -0.3035630 -0.4598280 -0.268287 63.50000 185.4003 54.83455
#> 6  0.000729417 -0.1666270  0.0428249 -0.532617 97.00000 189.2003 73.05570
#>        BSA      AGE      AST      ALT      BILI     CRCL      BMI      SEX
#> 1 1.889940 39.00003 29.62327 25.00005 23.946010  95.3004 21.80003 0.999999
#> 2 1.639042 42.00003 27.99996 30.24826  9.344359  95.5004 18.80000 0.999994
#> 3 1.853286 25.00000 24.40934 19.00005 10.262002 124.6026 26.10003 1.999994
#> 4 1.989243 39.00003 29.99996 24.00004 25.656010 130.7004 28.90004 0.999994
#> 5 1.808417 25.00000 22.99996 18.00000 11.972990 104.7004 18.50000 1.000003
#> 6 2.257851 41.00003 63.81090 82.00000 32.498010 104.4004 27.10003 0.999999
#>    RACEL_3   RACEL_2     NCIL_2    NCIL_1    GENO2    ETHNIC     SMOK
#> 1 -3.0e-07  0.999994 -0.0000041  0.999997 1.999991  0.103375 -2.8e-06
#> 2 -4.0e-06 -0.000003 -0.0000017  0.000000 1.680606  0.852023  8.3e-06
#> 3 -4.9e-06  1.000008  0.0000003  0.000005 1.999995 -0.000002  5.7e-06
#> 4 -4.6e-06  0.000002  1.0000019  0.000005 1.780071  0.999996 -1.7e-06
#> 5 -2.6e-06  0.000001 -0.0000020  0.000005 1.999994  0.000002 -2.0e-07
#> 6 -4.3e-06  0.999990  0.9999849 -0.000005 2.065208 -0.000005 -2.5e-06
#>     EBE_ETA1     EBE_ETA2   EBE_ETA3    EBE_ETA4  EBE_ETA5 WT_MISSING
#> 1  0.3118690  1.14713e-04  0.3038300 -0.17201700 -0.433899          0
#> 2  0.1272970  8.17392e-05 -0.0732683  0.00859341  0.206066          0
#> 3  0.1438600 -1.52807e-03 -0.0372967 -0.12800600 -0.184257          0
#> 4 -0.4356520  3.80416e-04 -0.2535530  0.23954400  0.294937          0
#> 5  0.0759199  1.13067e-03 -0.2450010 -0.12903100 -0.156757          0
#> 6 -0.3020010  6.39646e-04  0.1967590  0.02875560 -0.417324          0
#>   HT_MISSING LBWT_MISSING BSA_MISSING AGE_MISSING AST_MISSING ALT_MISSING
#> 1          0            0           0           0           1           0
#> 2          0            0           0           0           0           1
#> 3          0            0           0           0           1           0
#> 4          0            0           0           0           0           0
#> 5          0            0           0           0           0           0
#> 6          0            0           0           0           1           0
#>   BILI_MISSING CRCL_MISSING BMI_MISSING SEX_MISSING RACEL_3_MISSING
#> 1            0            0           0           0               0
#> 2            1            0           0           0               0
#> 3            0            1           0           0               0
#> 4            0            0           0           0               0
#> 5            0            0           0           0               0
#> 6            0            0           0           0               0
#>   RACEL_2_MISSING NCIL_2_MISSING NCIL_1_MISSING GENO2_MISSING ETHNIC_MISSING
#> 1               0              0              0             0              1
#> 2               0              0              0             1              1
#> 3               0              0              0             0              0
#> 4               0              0              0             1              0
#> 5               0              0              0             0              0
#> 6               0              0              0             1              0
#>   SMOK_MISSING
#> 1            0
#> 2            0
#> 3            0
#> 4            0
#> 5            0
#> 6            0